// src/types/netPricePublicIncome.ts
export interface NetPricePublicIncome {
  unitid: string;
  income_0_30000: number | null;
  income_0_48000: number | null;
  income_30001_48000: number | null;
  income_30001_75000: number | null;
  income_48001_75000: number | null;
  income_75001_110000: number | null;
  income_75000_plus: number | null;
  income_110001_plus: number | null;
}